misc file/class renames and tidy-ups
