+++ Headless.Core/Attributes/SupportedTargetsAttribute.cs
@@ -2,14 +2,13 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Headless.Core.Attributes
-{
-    [AttributeUsage(AttributeTargets.Class)]
-    public class SupportedTargetsAttribute(string name, string versions = "latest", string runtimes = "any") : Attribute
-    {
-        public const char KeyDelimiter = ';';
-        public const char PropertyDelimiter = '|';
+namespace Headless.Core.Attributes;
 
-        public IEnumerable<string> Keys { get; } = versions.Split(PropertyDelimiter).SelectMany(version => runtimes.Split(PropertyDelimiter).Select(runtime => $"{name}{KeyDelimiter}{version}{KeyDelimiter}{runtime}"));
-    }
+[AttributeUsage(AttributeTargets.Class)]
+public class SupportedTargetsAttribute(string name, string versions = "latest", string runtimes = "any") : Attribute
+{
+    public const char KeyDelimiter = ';';
+    public const char PropertyDelimiter = '|';
+
+    public IEnumerable<string> Keys { get; } = versions.Split(PropertyDelimiter).SelectMany(version => runtimes.Split(PropertyDelimiter).Select(runtime => $"{name}{KeyDelimiter}{version}{KeyDelimiter}{runtime}"));
 }
