--- Headless.Core/ICompileResult.cs
@@ -1,8 +1,0 @@
-﻿namespace Headless.Core
-{
-    public interface ICompileResult
-    {
-        bool IsSuccess { get; }
-        string Messages { get; }
-    }
-}
