--- Headless.Core/IReadScripts.cs
@@ -1,9 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace Headless.Core
-{
-    public interface IReadScripts
-    {
-        Task<ICompileResult> Compile(string script);
-    }
-}
