--- Headless.Core/IRunScripts.cs
@@ -1,10 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace Headless.Core
-{
-    public interface IRunScripts
-    {
-        Task<IInvocationResult> Run(ICompileResult compileResult);
-        Task<IInvocationResult<TResult>> Run<TResult>(ICompileResult compileResult);
-    }
-}
