+++ Headless.Core/IScriptCompiler.cs
@@ -1,0 +1,8 @@
+﻿using System.Threading.Tasks;
+
+namespace Headless.Core;
+
+public interface IScriptCompiler
+{
+    Task<ICompileResult> Compile(string script);
+}
